backend/requirements.txt
fastapi>=0.111.1,<0.112.0
tortoise-orm>=0.21.4,<1.0.0
pydantic>=2.8.2,<3.0.0
uvicorn>=0.30.1,<1.0.0
asyncpg>=0.29.0,<1.0.0
websockets>=12.0.0,<13.0.0
PyJWT>=2.8.0,<3.0.0
ipython>=8.26.0,<9.0.0